+++ src/main.js
@@ -4,8 +4,7 @@
 import DefaultLayout from '~/layouts/Default.vue'
 import VueScrollTo from 'vue-scrollto'
 import VueFuse from 'vue-fuse'
-import { VueHammer } from 'vue2-hammer'
-
+import Vue2TouchEvents from 'vue2-touch-events'
 
 export default function (Vue, { router, head, isClient }) {
   // Set default layout as a global component
@@ -18,7 +17,7 @@ export default function (Vue, { router, head, isClient }) {
 
   Vue.use(VueFuse)
 
-  Vue.use(VueHammer)
+  Vue.use(Vue2TouchEvents)
 
   head.meta.push({
     name: 'keywords',